[add] move transaction, added print message, add more fixes
