+++ src/momentum-safe/momentum-safe.ts
@@ -1,7 +1,6 @@
 import * as Aptos from "../web3/global";
 import {HexString, TxnBuilderTypes, BCS} from 'aptos';
-import {AptosCoinTransferTxnBuilder, AptosEntryTxnBuilder} from '../web3/transaction';
-import {Transaction} from "../web3/types";
+import {AptosCoinTransferTxnBuilder, AptosEntryTxnBuilder, Transaction} from '../web3/transaction';
 import {Account} from '../web3/account';
 import {
   vector,
