+++ src/web3/transaction.ts
@@ -1,5 +1,5 @@
-import {BCS, HexString, TxnBuilderTypes} from "aptos";
-import {Transaction} from "./types";
+import {BCS, HexString, TransactionBuilder, TxnBuilderTypes} from "aptos";
+import {Buffer} from "buffer/";
 
 const COIN_MODULE = "0x1::coin";
 const TRANSFER_METHOD = "transfer";
@@ -197,5 +197,23 @@ export class AptosEntryTxnBuilder extends AptosTxnBuilder {
         this._args,
       ),
     );
+  }
+}
+
+
+export class Transaction {
+  raw: TxnBuilderTypes.RawTransaction;
+
+  constructor(raw: TxnBuilderTypes.RawTransaction) {
+    this.raw = raw;
+  }
+
+  static deserialize(rawTx: Buffer): Transaction {
+    const deserializer = new BCS.Deserializer(rawTx.slice(32)); // skip prefix, see TransactionBuilder.getSigningMessage
+    return new Transaction(TxnBuilderTypes.RawTransaction.deserialize(deserializer));
+  }
+
+  getSigningMessage() {
+    return TransactionBuilder.getSigningMessage(this.raw);
   }
 }
